pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "google-cloud-bigquery==3.25.0",
    "google-cloud-core==2.4.1",
    "pydantic==2.6.4",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sqlalchemy==2.0.30",
    "uvicorn[standard]==0.29.0",
]
